Add waitFor + waitForAsync to Process
