+++ README.md
@@ -3,6 +3,7 @@
 [![badge-latest-release]][url-latest-release]
 
 [![badge-kotlin]][url-kotlin]
+[![badge-coroutines]][url-coroutines]
 [![badge-immutable]][url-immutable]
 
 ![badge-platform-android]
@@ -76,6 +77,7 @@ p.sigkill()
 
 <!-- TAG_DEPENDENCIES -->
 [badge-kotlin]: https://img.shields.io/badge/kotlin-1.9.21-blue.svg?logo=kotlin
+[badge-coroutines]: https://img.shields.io/badge/coroutines-1.7.3-blue.svg?logo=kotlin
 [badge-immutable]: https://img.shields.io/badge/immutable-0.1.0-blue.svg?style=flat
 
 <!-- TAG_PLATFORMS -->
@@ -98,5 +100,6 @@ p.sigkill()
 [url-latest-release]: https://github.com/05nelsonm/process/releases/latest
 [url-license]: https://www.apache.org/licenses/LICENSE-2.0
 [url-kotlin]: https://kotlinlang.org
+[url-coroutines]: https://github.com/Kotlin/kotlinx.coroutines
 [url-immutable]: https://github.com/05nelsonm/immutable
 [url-posix-spawn]: https://man7.org/linux/man-pages/man3/posix_spawn.3.html
