+++ library/process/api/process.api
@@ -7,6 +7,8 @@ public abstract class io/matthewnelson/process/Process : java/lang/Process {
 	public final fun isProcessAlive ()Z
 	public abstract fun sigkill ()Lio/matthewnelson/process/Process;
 	public abstract fun sigterm ()Lio/matthewnelson/process/Process;
+	public final fun waitFor-LRDsOJo (J)Ljava/lang/Integer;
+	public final fun waitForAsync-VtjQ1oo (JLkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
 public final class io/matthewnelson/process/Process$Builder {
@@ -24,9 +26,5 @@ public final class io/matthewnelson/process/ProcessException : java/lang/Runtime
 	public fun <init> (Ljava/lang/String;)V
 	public fun <init> (Ljava/lang/String;Ljava/lang/Throwable;)V
 	public fun <init> (Ljava/lang/Throwable;)V
-}
-
-public final class io/matthewnelson/process/ProcessExt {
-	public static final fun waitFor-HG0u8IE (Lio/matthewnelson/process/Process;J)Ljava/lang/Integer;
 }
