+++ library/process/build.gradle.kts
@@ -28,6 +28,7 @@ kmpConfiguration {
             sourceSetMain {
                 dependencies {
                     implementation(libs.immutable.collections)
+                    implementation(libs.kotlinx.coroutines.core)
                 }
             }
             sourceSetTest {
@@ -39,24 +40,6 @@ kmpConfiguration {
         }
 
         kotlin {
-            with(sourceSets) {
-                val jvmMain = findByName("jvmMain")
-                val nativeMain = findByName("nativeMain")
-
-                if (jvmMain != null || nativeMain != null) {
-                    val nonJsMain = maybeCreate("nonJsMain")
-                    val nonJsTest = maybeCreate("nonJsTest")
-
-                    nonJsMain.dependsOn(getByName("commonMain"))
-                    nonJsTest.dependsOn(getByName("commonTest"))
-
-                    jvmMain?.apply { dependsOn(nonJsMain) }
-                    findByName("jvmTest")?.apply { dependsOn(nonJsTest) }
-                    nativeMain?.apply { dependsOn(nonJsMain) }
-                    findByName("nativeTest")?.apply { dependsOn(nonJsTest) }
-                }
-            }
-
             targets.filterIsInstance<KotlinNativeTarget>().spawnCInterop()
         }
     }
