+++ library/process/src/jvmMain/kotlin/io/matthewnelson/process/Process.kt
@@ -13,17 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING", "ACTUAL_ANNOTATIONS_NOT_MATCH_EXPECT")
 
 package io.matthewnelson.process
 
 import io.matthewnelson.immutable.collections.toImmutableList
 import io.matthewnelson.immutable.collections.toImmutableMap
+import io.matthewnelson.process.internal.*
 import io.matthewnelson.process.internal.commonArg
 import io.matthewnelson.process.internal.commonEnvironment
 import io.matthewnelson.process.internal.commonIsAlive
 import io.matthewnelson.process.internal.commonWithEnvironment
+import kotlinx.coroutines.delay
 import java.io.IOException
+import kotlin.time.Duration
 
 /**
  * A Process.
@@ -54,6 +57,30 @@ public actual sealed class Process actual constructor(
     // w/o conflicting with java.lang.Process' function.
     @get:JvmName("isProcessAlive")
     public actual val isAlive: Boolean get() = commonIsAlive()
+
+    /**
+     * Blocks the current thread for the specified [timeout],
+     * or until [Process.exitCode] is available (i.e. the
+     * [Process] completed).
+     *
+     * @param [timeout] the [Duration] to wait
+     * @return The [Process.exitCode], or null if [timeout] is exceeded
+     * */
+    public actual fun waitFor(timeout: Duration): Int? {
+        return commonWaitFor(timeout) { Thread.sleep(it.inWholeMilliseconds) }
+    }
+
+    /**
+     * Delays the current coroutine for the specified [timeout],
+     * or until [Process.exitCode] is available (i.e. the
+     * [Process] completed).
+     *
+     * @param [timeout] the [Duration] to wait
+     * @return The [Process.exitCode], or null if [timeout] is exceeded
+     * */
+    public actual suspend fun waitForAsync(timeout: Duration): Int? {
+        return commonWaitFor(timeout) { delay(it) }
+    }
 
     /**
      * Kills the [Process] via signal SIGTERM and closes
