+++ library/process/src/nativeTest/kotlin/io/matthewnelson/process/NativeProcessUnitTest.kt
@@ -2,7 +2,7 @@ package io.matthewnelson.process
 
 import kotlin.experimental.ExperimentalNativeApi
 
-class NativeProcessUnitTest: ProcessNonJsBaseTest() {
+class NativeProcessUnitTest: ProcessBaseTest() {
 
     @OptIn(ExperimentalNativeApi::class)
     override val isUnixDesktop: Boolean = when (Platform.osFamily) {
@@ -10,4 +10,5 @@ class NativeProcessUnitTest: ProcessNonJsBaseTest() {
         OsFamily.LINUX -> true
         else -> false
     }
+    override val isNodeJS: Boolean = false
 }
